pull up clear jdt index (not used anymore)
