+++ tests/org.eclipse.emf.parsley.tests.swtbot/src/org/eclipse/emf/parsley/tests/swtbot/EmfParsleyDslUiAbstractTests.java
@@ -10,12 +10,7 @@
  */
 package org.eclipse.emf.parsley.tests.swtbot;
 
-import java.io.IOException;
-
-import org.apache.commons.io.FileUtils;
-import org.eclipse.jdt.core.JavaCore;
 import org.eclipse.swtbot.swt.finder.junit.SWTBotJunit4ClassRunner;
-import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
 import org.junit.runner.RunWith;
 
 @RunWith(SWTBotJunit4ClassRunner.class)
@@ -23,35 +18,6 @@ public abstract class EmfParsleyDslUiAbstractTests extends EmfParsleySWTBotAbstr
 	protected final String TEST_PROJ_NAME = "my.emfparsley.proj";
 
 	protected final String TEST_PROJ_NAME_BAD1 = "My-EmfParsley--project";
-
-	@Override
-	public void runAfterEveryTest() throws Exception {
-		super.runAfterEveryTest();
-//		clearJdtIndex();
-	}
-
-	protected void clearJdtIndex() {
-		var jdtMetadata = JavaCore.getPlugin().getStateLocation().toFile();
-		bot.waitUntil(new DefaultCondition() {
-			
-			@Override
-			public boolean test() {
-				System.err.println("Clean up index " + jdtMetadata.getAbsolutePath());
-				try {
-					FileUtils.deleteDirectory(jdtMetadata);
-				} catch (IOException e) {
-					System.err.println("retrying due to exception while cleaning");
-					return false;
-				}
-				return true;
-			}
-			
-			@Override
-			public String getFailureMessage() {
-				return "cannot delete " + jdtMetadata;
-			}
-		});
-	}
 
 	public void createDslProjectWithWizard() throws Exception {
 		createProjectWithoutTemplateInWorkspace(EMF_PARSLEY_CATEGORY,
